+++ golem-phaser/src/index.ts
@@ -1,8 +1,20 @@
 export { GameScene } from "./scene.js";
-export { createTiledLayer, loadTiledMap } from "./tiled.js";
+export {
+  createTiledLayer,
+  loadTiledMap,
+  loadTiledWorld,
+  refreshGpuTiledLayer,
+} from "./tiled.js";
 export type {
   CreateTiledLayerOptions,
+  LoadTiledWorldOptions,
+  MountedTiledWorld,
   TiledLayer,
   TiledLayerRenderMode,
   TiledLayerTilesets,
+  TiledMapSource,
+  TiledTilesetAsset,
+  TiledTilesetAssets,
+  TiledWorldData,
+  TiledWorldLayerOptions,
 } from "./tiled.js";